[FLOW-CORE] Cod de entitidade alterado de Serializable e Long de volta para Integer para evitar problemas com o Alocpro. Refactoring de alguns metodos.

+++ flow/persistence/src/main/java/br/net/mirante/singular/persistence/entity/RoleInstance.java
@@ -27,7 +27,7 @@ public class RoleInstance implements IEntityRole {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "CO_INSTANCIA_PAPEL")
-    private Long cod;
+    private Integer cod;
 
     @Temporal(TemporalType.TIMESTAMP)
     @Column(name = "DT_CRIACAO", nullable = false)
@@ -57,11 +57,11 @@ public class RoleInstance implements IEntityRole {
     }
 
     @Override
-    public Long getCod() {
+    public Integer getCod() {
         return cod;
     }
 
-    public void setCod(Long cod) {
+    public void setCod(Integer cod) {
         this.cod = cod;
     }
