+++ flow/persistence/src/main/java/br/net/mirante/singular/persistence/entity/Transition.java
@@ -1,9 +1,5 @@
 package br.net.mirante.singular.persistence.entity;
 
-import br.net.mirante.singular.flow.core.entity.IEntityTaskTransition;
-import br.net.mirante.singular.flow.core.entity.TransitionType;
-import br.net.mirante.singular.persistence.util.Constants;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -13,8 +9,11 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.NamedQuery;
 import javax.persistence.Table;
+
+import br.net.mirante.singular.flow.core.entity.IEntityTaskTransition;
+import br.net.mirante.singular.flow.core.entity.TransitionType;
+import br.net.mirante.singular.persistence.util.Constants;
 
 
 /**
@@ -28,7 +27,7 @@ public class Transition implements IEntityTaskTransition {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "CO_TRANSICAO")
-    private Long cod;
+    private Integer cod;
 
     @Column(name = "NO_TRANSICAO", nullable = false)
     private String name;
@@ -54,11 +53,11 @@ public class Transition implements IEntityTaskTransition {
     }
 
     @Override
-    public Long getCod() {
+    public Integer getCod() {
         return cod;
     }
 
-    public void setCod(Long cod) {
+    public void setCod(Integer cod) {
         this.cod = cod;
     }
